% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_living_meta.R
\name{create_living_analysis}
\alias{create_living_analysis}
\title{Create a template website for living meta-analysis}
\usage{
create_living_analysis(
  path = "living_analysis",
  title = "Living Meta-Analysis",
  authors = "Me",
  description = "Description of Living Analysis",
  output_dir = "docs",
  license = "CC-BY",
  google_analytics = "",
  margin_header = "",
  footer = paste0(license, " (", format(Sys.Date(), "\%Y"), ") ", authors),
  light_theme = "flatly",
  dark_theme = "darkly",
  df_print = "kable",
  open = rlang::is_interactive(),
  render = TRUE
)
}
\arguments{
\item{path}{path to the location for your website on your machine}

\item{title}{title of the living meta-analysis}

\item{authors}{authors of the living meta-analysis}

\item{description}{short description of the living meta-analysis}

\item{output_dir}{directory to render the website output to}

\item{license}{what license to add (CC-BY only for now)}

\item{google_analytics}{the google analytics address}

\item{margin_header}{defaults to ""}

\item{footer}{defaults to "license (YEAR) author"}

\item{light_theme}{"flatly" (see https://quarto.org/docs/output-formats/html-themes.html)}

\item{dark_theme}{"darkly"}

\item{df_print}{how to print tables (default, kable, tibble, paged)}

\item{open}{whether to activate the new project in RStudio}

\item{render}{whether to render the quarto website when opening}
}
\value{
sets up a project and renders the demo website
}
\description{
#' Create a template website for a living meta-analysis. You can add details of the methods/protocol, results, and allow others to contribute
}
