% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frankensteinR-package.R
\docType{package}
\name{frankensteinR-package}
\alias{frankensteinR-package}
\alias{meta-package}
\alias{frankensteinR}
\title{frankensteinR}
\description{
Creates a template quarto website for a living meta-analysis.
}
\keyword{package}
